pyproject.toml
[tool.poetry]
name = "dls-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dls_project"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "1.24"
scipy = "^1.11.1"
tensorflow = "^2.13.0"
python-telegram-bot = "^20.3"
imageio = "^2.31.1"
opencv-python = "^4.8.0.74"
python-dotenv = "^1.0.0"
matplotlib = "^3.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
